#8 Criação de serviço para envio de email
