+++ app/DI/ServicesConfig.cs
@@ -8,6 +8,7 @@ namespace app.DI
         public static void AddConfigServices(this IServiceCollection services, IConfiguration configuration)
         {
             services.AddScoped<IUsuarioService, UsuarioService>();
+            services.AddScoped<IEmailService, EmailService>();
         }
     }
 }
