+++ service/Interfaces/IEmailService.cs
@@ -1,0 +1,7 @@
+﻿namespace service.Interfaces
+{
+    public interface IEmailService
+    {
+        public void EnviarEmail(string emailDestinatario, string assunto, string corpo);
+    }
+}
